changfed int to long

+++ bit_manipulation/1-print_binary.c
@@ -9,7 +9,7 @@
 
 void print_binary(unsigned long int n)
 {
-unsigned int i = n;
+unsigned long i = n;
 
 if (i > 1)
 {
